move style logic, rename and move files to main
