+++ src/lib/styles/ExpertBlock.styles.ts
@@ -1,0 +1,9 @@
+import { makeStyles } from '@material-ui/core';
+
+export const useStyles = makeStyles(() => ({
+  photo: {
+    width: '100%',
+    height: '100%',
+    objectFit: 'cover',
+  },
+}));
